@@ -9,7 +9,7 @@
   },
   "license": "MIT",
   "engines": {
-    "vscode": "^1.33.0"
+    "vscode": "^1.35.0"
   },
   "scripts": {
     "nx": "nx",
@@ -82,7 +82,7 @@
     "@nrwl/angular": "8.0.0",
     "core-js": "2.5.7",
     "ij-rpc-client": "^0.3.2",
-    "node-pty-prebuilt": "0.7.6",
+    "node-pty": "^0.8.1",
     "source-map-explorer": "^1.8.0",
     "stylelint": "^9.10.1",
     "stylelint-config-standard": "^18.2.0",
@@ -97,7 +97,7 @@
     "@angular/common": "8.0.0",
     "@angular/compiler": "8.0.0",
     "@angular/compiler-cli": "8.0.0",
-    "@angular/flex-layout": "7.0.0-beta.24",
+    "@angular/flex-layout": "8.0.0-beta.26",
     "@angular/forms": "8.0.0",
     "@angular/language-service": "8.0.0",
     "@angular/material": "8.0.0",
@@ -118,7 +118,7 @@
     "@types/d3-selection": "^1.4.1",
     "@types/d3-shape": "^1.3.1",
     "@types/d3-transition": "^1.1.4",
-    "@types/electron-store": "^1.3.0",
+    "@types/electron-store": "^3.2.0",
     "@types/fontfaceobserver": "^0.0.6",
     "@types/get-port": "^4.0.1",
     "@types/graphql": "14.0.5",
@@ -133,33 +133,33 @@
     "@types/strip-json-comments": "^0.0.30",
     "@types/tmp": "^0.0.33",
     "@types/universal-analytics": "0.4.2",
-    "apollo-angular": "1.5.0",
-    "apollo-angular-link-http": "1.5.0",
-    "apollo-cache-inmemory": "1.4.2",
-    "apollo-client": "^2.4.8",
-    "apollo-link": "1.2.8",
-    "apollo-link-error": "1.1.7",
-    "apollo-server-express": "^2.4.0",
-    "codelyzer": "4.5.0",
+    "apollo-angular": "1.6.0",
+    "apollo-angular-link-http": "1.8.0",
+    "apollo-cache-inmemory": "1.6.2",
+    "apollo-client": "^2.6.3",
+    "apollo-link": "1.2.12",
+    "apollo-link-error": "1.1.11",
+    "apollo-server-express": "^2.6.9",
+    "codelyzer": "5.1.0",
     "cypress": "^3.1.5",
     "d3-hierarchy": "^1.1.8",
     "d3-selection": "^1.4.0",
     "d3-shape": "^1.3.4",
     "d3-transition": "^1.2.0",
     "dotenv": "6.2.0",
-    "electron": "2.0.14",
-    "electron-builder": "20.28.4",
-    "electron-installer-dmg": "2.0.0",
-    "electron-packager": "13.0.1",
-    "electron-rebuild": "^1.8.4",
-    "electron-store": "2.0.0",
-    "electron-updater": "3.1.2",
+    "electron": "4.2.6",
+    "electron-builder": "21.0.15",
+    "electron-installer-dmg": "3.0.0",
+    "electron-packager": "14.0.1",
+    "electron-rebuild": "^1.8.5",
+    "electron-store": "3.3.0",
+    "electron-updater": "4.1.2",
     "express": "4.16.4",
     "fast-glob": "^2.2.6",
     "fix-path": "2.1.0",
     "fontfaceobserver": "^2.1.0",
     "get-port": "4.1.0",
-    "graphql": "14.1.1",
+    "graphql": "14.2.1",
     "graphql-code-generator": "0.16.1",
     "graphql-codegen-typescript-apollo-angular": "0.16.1",
     "graphql-codegen-typescript-client": "0.16.1",
@@ -169,12 +169,11 @@
     "graphql-tag": "2.10.1",
     "graphql-tools": "4.0.4",
     "husky": "^1.3.1",
-    "jasmine-marbles": "0.5.0",
+    "jasmine-marbles": "^0.6.0",
     "jest": "24.1.0",
     "jest-jasmine2": "23.6.0",
     "jest-preset-angular": "7.0.0",
     "ng-packagr": "5.2.0",
-    "node-pty-prebuilt": "0.7.6",
     "nps": "5.9.3",
     "nps-utils": "1.7.0",
     "opn": "^5.4.0",
@@ -189,9 +188,9 @@
     "shelljs": "0.8.3",
     "shx": "^0.3.2",
     "strip-json-comments": "2.0.1",
-    "ts-jest": "23.10.5",
+    "ts-jest": "24.0.2",
     "ts-node": "8.0.2",
-    "tsickle": "0.34.3",
+    "tsickle": "0.35.0",
     "tslib": "1.9.3",
     "tslint": "5.12.1",
     "tslint-no-unused-expression-chai": "^0.1.4",
@@ -199,13 +198,10 @@
     "universal-analytics": "0.4.20",
     "uuid": "3.3.2",
     "vsce": "^1.65.0",
-    "vscode": "^1.1.33",
-    "xterm": "3.11.0",
+    "vscode": "^1.1.35",
+    "xterm": "3.14.5",
     "zlib": "^1.0.5",
     "zone.js": "0.9.1"
-  },
-  "resolutions": {
-    "webpack": "4.29.0"
   },
   "jest": {
     "moduleFileExtensions": [
