@@ -9,13 +9,13 @@ import {
   ViewChild
 } from '@angular/core';
 import { Title } from '@angular/platform-browser';
-import { NavigationEnd, Router, RouterOutlet } from '@angular/router';
+import { RouterOutlet } from '@angular/router';
 import {
   Breadcrumb,
   ContextualActionBarService
 } from '@nrwl/angular-console-enterprise-frontend';
 import { Observable } from 'rxjs';
-import { filter, map } from 'rxjs/operators';
+import { map } from 'rxjs/operators';
 
 const TITLE_SEPARATOR = ' | ';
 
@@ -49,26 +49,9 @@ export class AppComponent implements OnInit, OnDestroy {
   constructor(
     readonly settings: Settings,
     private readonly contextualActionBarService: ContextualActionBarService,
-    private readonly titleService: Title,
-    router: Router
+    private readonly titleService: Title
   ) {
     settings.fetch().subscribe();
-
-    router.events
-      .pipe(filter((e): e is NavigationEnd => e instanceof NavigationEnd))
-      .subscribe(e => {
-        if (e.urlAfterRedirects === '/connect/support') {
-          // TODO: This logic belong within angular-console-enterprise-frontend
-          this.contextualActionBarService.breadcrumbs$.next([
-            { title: 'Connect' }
-          ]);
-        } else if (e.urlAfterRedirects === '/settings') {
-          // TODO: This logic belong within the settings component after electron redesign.
-          this.contextualActionBarService.breadcrumbs$.next([
-            { title: 'Settings' }
-          ]);
-        }
-      });
   }
 
   private readonly titleSubscription = this.contextualActionBarService.breadcrumbs$
