@@ -1,13 +1,13 @@
 import { Injectable } from '@angular/core';
+import { BehaviorSubject } from 'rxjs';
 import { first, tap } from 'rxjs/operators';
 
 import {
+  Maybe,
   Settings as SettingsModels,
   SettingsGQL,
-  UpdateSettingsGQL,
-  Maybe
+  UpdateSettingsGQL
 } from './generated/graphql';
-import { BehaviorSubject } from 'rxjs';
 
 export { Settings as SettingsModels } from './generated/graphql';
 
@@ -30,6 +30,7 @@ export class Settings {
     channel: 'latest',
     disableAnimations: true,
     isWsl: false,
+    isWindows: false,
     useNvm: false
   };
 
@@ -107,8 +108,8 @@ export class Settings {
     return this.settings.enableDetailedStatus;
   }
 
-  getChannel() {
-    return this.settings.channel;
+  isWindows() {
+    return this.settings.isWindows;
   }
 
   private readonly disabledAnimationsSubject = new BehaviorSubject<
